+++ lib/cliutils/prefs/pref_validation.rb
@@ -8,7 +8,8 @@ module CLIUtils
     # and a result message.
     Validator = Struct.new(:code, :message)
 
-    # Validates that a value is only letters.
+    # Validates that a value is only letters
+    # and spaces.
     # @param [String] text The text to inspect
     # @return [Boolean]
     def self.alphabetic(text)
@@ -17,7 +18,8 @@ module CLIUtils
       Validator.new(c, m)
     end
 
-    # Validates that a value is only letters and numbers.
+    # Validates that a value is only letters, numbers
+    # and spaces.
     # @param [String] text The text to inspect
     # @return [Boolean]
     def self.alphanumeric(text)
@@ -35,6 +37,15 @@ module CLIUtils
       Validator.new(c, m)
     end
 
+    # Validates that a value is a datetime.
+    # @param [String] text The text to inspect
+    # @return [Boolean]
+    def self.datetime(text)
+      m = "Response is not a datetime: #{ text }"
+      c = !(DateTime.parse(text) rescue nil).nil?
+      Validator.new(c, m)
+    end
+
     # Validates that a value is passed and is not
     # empty.
     # @param [String] text The text to inspect
@@ -48,9 +59,28 @@ module CLIUtils
     # Validates that a value is some sort of number.
     # @param [String] text The text to inspect
     # @return [Boolean]
-    def self.numeric(text)
+    def self.number(text)
       m = "Response is not a number: #{ text }"
       c = text.to_s =~ /\A[-+]?\d*\.?\d+\z/
+      Validator.new(c, m)
+    end
+
+    # Validates that a filepath exists on the
+    # local filesystem.
+    # @param [String] text The text to inspect
+    # @return [Boolean]
+    def self.filepath_exists(text)
+      m = "Path does not exist locally: #{ text }"
+      c = Pathname.new(text).exist?
+      Validator.new(c, m)
+    end
+
+    # Validates that a value is a time.
+    # @param [String] text The text to inspect
+    # @return [Boolean]
+    def self.time(text)
+      m = "Response is not a time: #{ text }"
+      c = !(Time.parse(text) rescue nil).nil?
       Validator.new(c, m)
     end